requirements.txt
dash==2.8.1
mediapipe==0.9.1.0
numpy==1.22.4
opencv_contrib_python==4.7.0.68
Pillow==9.4.0
